src/app/configs/configs.service.ts
/*********************************************************************
* Copyright (c) Intel Corporation 2022
* SPDX-License-Identifier: Apache-2.0
**********************************************************************/

import { HttpClient } from '@angular/common/http'
import { Injectable } from '@angular/core'
import { Observable, throwError } from 'rxjs'
import { catchError } from 'rxjs/operators'
import { environment } from 'src/environments/environment'
import { PageEventOptions } from 'src/models/models'
import { AuthService } from '../auth.service'
import { Config, ConfigsResponse } from './configs.constants'

@Injectable({
  providedIn: 'root'
})
export class ConfigsService {
  private readonly url = `${environment.rpsServer}/api/v1/admin/ciraconfigs`

  constructor (private readonly authService: AuthService, private readonly http: HttpClient) {

  }

  getData (pageEvent?: PageEventOptions): Observable<ConfigsResponse> {
    let query = this.url
    if (pageEvent) {
      query += `?$top=${pageEvent.pageSize}&$skip=${pageEvent.startsFrom}&$count=${pageEvent.count}`
    } else {
      query += '?$count=true'
    }
    return this.http.get<ConfigsResponse>(query)
      .pipe(
        catchError((err) => {
          const errorMessages = this.authService.onError(err)
          return throwError(errorMessages)
        })
      )
  }

  getRecord (name: string): Observable<Config> {
    return this.http.get<Config>(`${this.url}/${name}`)
      .pipe(
        catchError((err) => {
          const errorMessages = this.authService.onError(err)
          return throwError(errorMessages)
        })
      )
  }

  update (ciraConfig: Config): Observable<Config> {
    return this.http.patch<Config>(this.url, ciraConfig)
      .pipe(
        catchError((err) => {
          const errorMessages = this.authService.onError(err)
          return throwError(errorMessages)
        })
      )
  }

  create (ciraConfig: Config): Observable<Config> {
    return this.http.post<Config>(this.url, ciraConfig)
      .pipe(
        catchError((err) => {
          const errorMessages = this.authService.onError(err)
          return throwError(errorMessages)
        })
      )
  }

  loadMPSRootCert (): Observable<any> {
    // ToDo: Need to pass the mps server address to get the certs for each specific mps server
    const options = { responseType: 'text' } as any
    return this.http.get<string>(`${environment.mpsServer}/api/v1/ciracert`, options)
      .pipe(
        catchError((err) => {
          const errorMessages: string[] = []
          errorMessages.push('Error loading CIRA config')
          return throwError(err)
        })
      )
  }

  delete (name: string): Observable<any> {
    return this.http.delete(`${this.url}/${name}`)
      .pipe(
        catchError((err) => {
          const errorMessages = this.authService.onError(err)
          return throwError(errorMessages)
        })
      )
  }
}
